# packages/schemas/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "CitationTarget": {
        "additionalProperties": false,
        "properties": {
          "citation_id": {
            "format": "uuid",
            "title": "Citation Id",
            "type": "string"
          },
          "source_anchor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Anchor Id"
          },
          "source_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision Id"
          },
          "target_type": {
            "enum": [
              "source_anchor",
              "web_evidence"
            ],
            "title": "Target Type",
            "type": "string"
          },
          "web_evidence_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Web Evidence Id"
          }
        },
        "required": [
          "citation_id",
          "target_type"
        ],
        "title": "CitationTarget",
        "type": "object"
      },
...
      "CsrfResponse": {
        "additionalProperties": false,
        "properties": {
          "csrf_token": {
            "maxLength": 512,
            "minLength": 20,
            "title": "Csrf Token",
            "type": "string"
          }
        },
        "required": [
          "csrf_token"
        ],
        "title": "CsrfResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Frozen Phase 2 durable-core wire contract.",
    "title": "AXit Meeting RAG API",
    "version": "0.1.0-phase2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/citations/{citation_id}/resolve": {
      "get": {
        "operationId": "resolveCitation",
        "parameters": [
          {
            "in": "path",
            "name": "citation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Citation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationTarget"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Resolve Citation Contract"
      }
    },
    "/api/csrf": {
      "get": {
        "operationId": "getCsrf",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Get Csrf Contract"
      }
    },
...
  }
}